+++ components/inspectit-ocelot-configurationserver-ui/src/components/views/status/dialogs/AgentConfigurationDialog.js
@@ -1,0 +1,96 @@
+import React from 'react';
+import { Dialog } from 'primereact/dialog';
+import { connect } from 'react-redux';
+import { Button } from 'primereact/button';
+import SyntaxHighlighter from 'react-syntax-highlighter';
+import { tomorrowNightBlue } from 'react-syntax-highlighter/dist/cjs/styles/hljs';
+import axios from '../../../../lib/axios-api';
+
+class AgentConfigurationDialog extends React.Component {
+  state = {
+    configurationValue: '',
+  };
+
+  constructor(props) {
+    super(props);
+    this.getConfiguration(this.props.attributes);
+  }
+
+  getConfiguration = (attributes) => {
+    const requestParams = this.solveRegexValues(attributes);
+    if (!requestParams) {
+      return;
+    }
+
+    axios
+      .get('/configuration/agent-configuration', {
+        params: { ...requestParams },
+      })
+      .then((res) => {
+        this.setState({
+          configurationValue: res.data,
+        });
+      })
+      .catch(() => {
+        return 'error';
+      });
+  };
+
+  download = () => {
+    var blob = new Blob([this.state.configurationValue], { type: 'text/x-yaml' });
+    console.log(blob);
+    this.url = window.URL.createObjectURL(blob);
+    return this.url;
+  };
+
+  solveRegexValues = (obj = {}) => {
+    try {
+      let res = {};
+      Object.keys(obj).map((key) => {
+        const randexp = new RandExp(obj[key]);
+        randexp.max = randexp.min || 0;
+        res[key] = randexp.gen();
+      });
+      return res;
+    } catch {
+      this.props.showInfoMessage('Downloading Config File Failed', 'The given attribute values are not a regular expression');
+      return null;
+    }
+  };
+
+  render() {
+    return (
+      <Dialog
+        header={'Agent Configuration'}
+        modal={true}
+        visible={this.props.visible}
+        onHide={this.props.onHide}
+        footer={
+          <div>
+            <a href={this.download()} download="agent-config.yml">
+              <Button label="Download" className="p-button-primary" />
+            </a>
+            <Button label="Cancel" className="p-button-secondary" onClick={this.handleClose} />
+          </div>
+        }
+      >
+        <SyntaxHighlighter language="yaml" style={tomorrowNightBlue}>
+          {this.state.configurationValue}
+        </SyntaxHighlighter>
+      </Dialog>
+    );
+  }
+
+  /**
+   * Closing dialog.
+   */
+  handleClose = (success = true) => {
+    if (success) {
+      this.props.onHide();
+    }
+  };
+}
+
+const mapDispatchToProps = {};
+
+export default connect(null, mapDispatchToProps)(AgentConfigurationDialog);
